+++ internal/plugin/zz_provider.go
@@ -22,6 +22,7 @@ import (
 	"github.com/aiven/terraform-provider-aiven/internal/plugin/service/flink/application"
 	"github.com/aiven/terraform-provider-aiven/internal/plugin/service/flink/deployment"
 	"github.com/aiven/terraform-provider-aiven/internal/plugin/service/governance/access"
+	"github.com/aiven/terraform-provider-aiven/internal/plugin/service/kafka/topic"
 	"github.com/aiven/terraform-provider-aiven/internal/plugin/service/kafka/topiclist"
 	"github.com/aiven/terraform-provider-aiven/internal/plugin/service/kafkaschema/registryacl"
 	database1 "github.com/aiven/terraform-provider-aiven/internal/plugin/service/mysql/database"
@@ -65,6 +66,7 @@ func Resources() map[string]func() resource.Resource {
 		"aiven_flink_application_deployment":        adapter.NewLazyResource(deployment.ResourceOptions),
 		"aiven_governance_access":                   adapter.NewLazyResource(access.ResourceOptions),
 		"aiven_kafka_schema_registry_acl":           adapter.NewLazyResource(registryacl.ResourceOptions),
+		"aiven_kafka_topic":                         adapter.NewLazyResource(topic.ResourceOptions),
 		"aiven_mysql_database":                      adapter.NewLazyResource(database1.ResourceOptions),
 		"aiven_mysql_user":                          adapter.NewLazyResource(user1.ResourceOptions),
 		"aiven_opensearch_security_plugin_config":   adapter.NewLazyResource(securitypluginconfig.ResourceOptions),
@@ -96,6 +98,7 @@ func DataSources() map[string]func() datasource.DataSource {
 		"aiven_connection_pool":                     adapter.NewLazyDataSource(connectionpool.DataSourceOptions),
 		"aiven_flink_application":                   adapter.NewLazyDataSource(application.DataSourceOptions),
 		"aiven_kafka_schema_registry_acl":           adapter.NewLazyDataSource(registryacl.DataSourceOptions),
+		"aiven_kafka_topic":                         adapter.NewLazyDataSource(topic.DataSourceOptions),
 		"aiven_kafka_topic_list":                    adapter.NewLazyDataSource(topiclist.DataSourceOptions),
 		"aiven_mysql_database":                      adapter.NewLazyDataSource(database1.DataSourceOptions),
 		"aiven_mysql_user":                          adapter.NewLazyDataSource(user1.DataSourceOptions),